Move integration tests to new file
